refactor: 패키지명 (model) -> (domain)으로 변경
