+++ src/main/java/domain/Lotto.java
@@ -1,6 +1,6 @@
-package model;
+package domain;
 
-import static model.LottoInformation.*;
+import static domain.LottoInformation.*;
 import java.util.List;
 
 public class Lotto {
