+++ src/main/java/domain/LottoInformation.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 public class LottoInformation {
     public static final int LOTTO_COUNT = 6;
