+++ src/main/java/domain/LottoResult.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import enums.LottoRanking;
