+++ src/main/java/domain/LottoWinningChecker.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import enums.LottoRanking;
