+++ src/main/java/domain/Lottos.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.util.ArrayList;
 import java.util.List;
