+++ src/main/java/domain/Numbers.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.util.Comparator;
 import java.util.List;
