+++ src/main/java/domain/RandomGenerator.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.util.*;
