+++ src/main/java/domain/WinningNumberWithBonusNumber.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 public record WinningNumberWithBonusNumber(
         Lotto winningNumber,
