+++ src/test/java/LottoMachineTest.java
@@ -1,5 +1,5 @@
-import model.LottoMachine;
-import model.Price;
+import domain.LottoMachine;
+import domain.Price;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
