+++ src/test/java/LottoWinningCheckerTest.java
@@ -1,5 +1,5 @@
 import enums.LottoRanking;
-import model.*;
+import domain.*;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.params.ParameterizedTest;
