+++ src/test/java/PriceTest.java
@@ -1,4 +1,4 @@
-import model.Price;
+import domain.Price;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
@@ -22,5 +22,5 @@ class PriceTest {
         assertThatThrownBy(() -> new Price("asdasd"))
                 .isInstanceOf(IllegalArgumentException.class);
     }
-    
+
 }
