+++ src/test/java/RandomGeneratorTest.java
@@ -1,4 +1,4 @@
-import model.RandomGenerator;
+import domain.RandomGenerator;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
