+++ src/test/java/WinningNumberWithBonusNumberTest.java
@@ -1,6 +1,6 @@
-import model.Lotto;
-import model.Numbers;
-import model.WinningNumberWithBonusNumber;
+import domain.Lotto;
+import domain.Numbers;
+import domain.WinningNumberWithBonusNumber;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
